+++ src/heavyedge/api/__init__.py
@@ -2,6 +2,7 @@
 
 __all__ = [
     "preprocess",
+    "fill_after",
     "outlier",
     "mean",
     "landmarks_type2",
@@ -11,4 +12,4 @@ __all__ = [
 ]
 
 from .landmarks import landmarks_type2, landmarks_type3, plateau_type2, plateau_type3
-from .profile import mean, outlier, preprocess
+from .profile import fill_after, mean, outlier, preprocess